Back/requirements.txt
fastapi
uvicorn[standard]
numpy
python-multipart
httpx
pypdf2
python-docx
odfpy
